# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/carrito/{id}": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Obtener carrito por id",
        "description": "Retorna un carrito por id",
        "operationId": "obtenerCarritoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Se encontro el carrito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarritoResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No se encontro el carrito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Carrito"
        ],
        "summary": "Actualizar carrito",
        "description": "Actualiza un carrito",
        "operationId": "actualizarCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Introducir ID del carrito que se desea actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Carrito actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarritoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No se encontro el carrito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Eliminar carrito",
        "description": "Elimina un carrito",
        "operationId": "eliminarProductoDelCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Introducir ID del carrito que se desea eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Carrito eliminado",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No se encontro el carrito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/carrito": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Listar carritos",
        "description": "Retorna una lista de todos los carritos",
        "operationId": "listarCarrito",
        "responses": {
          "200": {
            "description": "Se encontraron carritos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarritoResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron carritos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Carrito"
        ],
        "summary": "Agregar producto al carrito",
        "description": "Agrega un producto al carrito",
        "operationId": "agregarProductoAlCarrito",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Producto agregado al carrito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarritoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Rol": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "usuarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Usuario"
            },
            "uniqueItems": true
          }
        }
      },
      "Usuario": {
        "type": "object",
        "description": "Introducir objeto de tipo Usuario a guardar",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rol"
            },
            "uniqueItems": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "mensaje": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "erroresValidacion": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "productos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Producto"
            }
          }
        }
      },
      "Producto": {
        "type": "object",
        "description": "Introducir objeto de tipo Producto a guardar",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        }
      },
...
      "Carrito": {
        "type": "object",
        "description": "Introducir objeto de tipo Carrito a guardar",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CarritoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "usuarioId": {
            "type": "integer",
            "format": "int64"
          },
          "productoId": {
            "type": "integer",
            "format": "int64"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}